Update open-source documentation
As an open-source community member, it's important to have first class
documentation so that any one feels like they can contribute to our
codebase. This adds a code of conduct, a contributing guide, and expands
the readmes to be more detailed and provide readmes for each crate.

Also, ran `cargo fmt`.