Setup for confirming the rustfmt
Remove testing for anything other than stable. We should desire to pass
on nightly and others but for now we can target stable. The reason is
that the rustfmt crate is not entirely stable so some functionality is
different on stable versus nightly. Also, add a rustfmt.toml to enforce
styles that we want in the future. It's empty right now.