Implement Sync Client
This commit implements the synchronous client library utilizing the
reqwest http client. It maps the end point into http and then converts
it to a reqwest request. This is then issued to the server and parsed
into the resulting response type correctly.

I've also divided up the client library into two components, the sync
and async module to allow easy choice between them. Generally the sync
will be useful for the CLI but async may be better for the actual web
server itself.

Lastly, I have written examples issuing a get query to the account
details end point :)