Add all operations to resources lib
This commit adds the 10 different operations to the resources lib.
Transactions in stellar are made up of a list of operations. Each operation is an
individual command that mutates the ledger.  These operations can all be
returned by the horizon API in polling or streaming mode.  Given that
the operations come from the same endpoints, they all must share a
deserializer.