Refactor operations details creation into a single match
To simplify how the matching works, I've refactored it into a single
match that errors on types that don't match their type. The key to
making this match work was changing the intermediate type field from a
String to a &'a str. This allows rust to perform a literal match as
opposed to a conditional one.

Another aspect of the change was to return early with an Err if a match
determines the error. This allows us to return an explicit operation
detail from the match and return from the function if it errors.