Rename for brevity
Some of the names are very long and redundant with the module that they
are contained in. By renaming them privately, the file itself can be a
bit less redundant while not changing the public interface at all. It's
pretty easy to assume that the `Intermediate` struct is representative
of the operation as it used in the operation module.

Similarly, `OperationDetail` is renamed to `OperationKind` and is aliased
to `Kind` privately.

I've also renamed `operation_type` to `kind` as it means the
same thing with many fewer characters.