DRY some of the code up
Two main changes:
1. Compare strings to str literals directly
2. Move all missing field matching to the last catch all.

The latter dries up much of the file and still provides the same amount
of information about what failed to the client developer. It also
removes a lot of noise from the file that I felt was making it difficult
to read.

The former is more efficient and easier to read.