Add all effect resources.
An effect is a change in the ledger that is created by an
operation. We collect these resources to make our application
eventually consistent with the state of the world in the stellar ledger.

This resource is modeled in the same way that operations are modeled
creating for consistency. This includes shortening
"EffectDetail" to "Kind", the simplified matching patterns, and the same
error handling patterns.