Fix the warnings from `cargo doc`
There are warnings in the rendering differences when running `cargo
doc`. This is because rust is switching from `hoedown` to `pulldown` as
a rendering client. The changes are minor but auto-links are no longer
correctly supported. If you want to put a link you need to wrap it with
`<` and `>` or put it in a standard markdown link: `[text](link)`.

Another minor difference that I added was putting an extra space above
raw links to separate them from the descriptive text a bit. I also
wrapped one of the links into text so that it reads more intuitively.

For testing this, you should be able to run `cargo doc` and see no more
warnings on this repo.