Implement bad response parsing for client errors
If the request to stellar is incorrect we may get back a bad response.
This response has a defined type and we can parse it from that json type
into our known StellarErrors. Along the way I noticed that the "type"
field is actually a url and not a slug at all. I've corrected it.

I also renamed the `ErrorType` field into `Kind` and began using the
`FromStr` trait to parse the `Kind` instead of a custom method. It might
be useful to parse this into a separate field as well to return the URL
to visit in the future as reference.