Implement Iterator for RecordIter
So, the actual Records struct can't really be an iterator across all
records due to it's lack of knowledge about the client and the endpoint
it's part of. But, we can create a record iter for use with the sync client!

This commit implements that type. It requires some traits to be
implemented by the resource and the endpoint for it to work correctly.
It needs to be able to clone the endpoint and the resource. Also, the
endpoint must have an associated type of a `Record`. This allows us to
paginate based on the endpoint, cache the returned records, then iterate
through them.

This immediately allows us to paginate across the entire set of records
associated with whatever endpoint we want. This _should_ be lazy based
on the way that rust iterators work themselves as they are lazy by
default.