Renames the trait EndPoint to IntoRequest
To follow along with what appear to be common naming conventions in rust
I'm proposing to rename the EndPoint trait to IntoRequest. This does not
change the module (I still like classifying it as endpoints at the
module lever) but I want to be explicit about what the trait does, not
what it's supposed to "be". Since the point of the trait is to consume
itself and produce a request, this trait name seems more appropriate.

see this issue in the rust-nursery:
https://github.com/rust-lang-nursery/api-guidelines/issues/28