Major enhancements to cli
Several things went into this change. Some was refactoring and some were
new enhancements.

*Added All Assets*

The first change was the most straight forward, the assets endpoint was
added. This end point features filtering via the command line args of
code and issuer.

*Switched to using the Iter*

The second change was the migration to using the iterator. This majorly
simplifies the code and eliminates the need for the recursion that was
there to begin with. I use the `enumerate` iterator to pair each line
with the index so that we can paginate.

*Functions return result types now*

Instead of duplicating the exception handling through out, instead we
just have each function return the client errors and handle it all at
the end.

*Choose the network!*

Another option that was added was the ability specify which network you
wished to join!

*Move the groups of functions to modules*

The file was feeling a little complex so I moved each group of commands
to their own module file. They were already modules so this was very
straightforward.