Add payments to the account endpoint
As a means to introspect the payments in which an account was sender or
receiver, this implements the account endpoint for payments.

In addition, this commit also improves the doc tests for other account
endpoint actions. By moving the repeated logic in each test to a single
top-level test for each action-associated Struct, we can reduce the time
spent making external calls and keep the function tests small and focused.

Fixes #78