Add an order trait and derive it
To simplify the creation of endpoints and maintenance, this creates an
`Order` trait and renames the direction of the order to `Direction`.
This allows us to then use a procedural macro to derive the trait. Also,
it enables some additional simplifications in the cli in the future.

Also in this change is modifying asset_code and asset_issuer setters to
use the `with_` prefix.

closes #135