Remove Serialize from resources
From all of the top level resources, the serialize implementation has
been removed. This will force users to implement their own
serializations and not be dependent on mimicing the same schema as the
horizon API. It does leave the serialize trait on some types that are
basal. These are `AssetIdentifier`, `Amount`, `PriceRatio`, and
`Base64String`. I felt that these would likely be reused in other types
and thus are useful to have serialize on.

closes #128