Implement Trade Aggregations Endpoint
To make the trade aggregations endpoint work in a way that can be
iterated against, we need to walk back some of the design considerations
around `Records`. The type records should actually end up just return
parsed Option<http::Uri> as the links. This can then be used in
conjunction with the newly introduced `TryIntoUri` trait. If an endpoint
can be converted into itself from a uri and it contains `Records` then
it can be iterated. For all iterated endpoints, we'll need to finish
implementing this trait.

Another exciting part about this change is that the `FlatRecords` type
can be retired as `Records` will now work without a next url.