Add trades endpoint to CLI
In the process also updated the name of the trade endpoint initializer
from with_trade_pair to with_asset_pair as that more accuately describes
what the method does.

Also created a new error enum in the CLI that wraps both the client
errors and introduces the concept of an input error.  All cli operations
updated to use the new CliError enum.