Abstraction for rendering
To simplify the rendering process I've created a way to implement the
render trait for a given resource. This can then be entirely inferred in
the cli commands. In theory, we could have `Simple`, `Color`, `CSV` as
various "types" to implement render on. Then if it's implemented the cli
will just work.