All effects added to CLI
Implements the effects in the CLI. This actually uncovered a couple of
other things that needed to be refactored to work correctly.

First, `Flags` needed to be made public. In the process of exposing it I
renamed it to `Flags` from `Flag` since it contains multiple flags.

Second, there are three undocumented effects that needed to be
implemented. These include data_created, data_updated, and data_removed.

Lastly, the effects were added as a render to the cli.