Add operations for account to CLI
Adds a command to request all operations relating to a given account
to the CLI application.

This also implements Render for each kind of operation, so the details
of each operation get included with the output. The formatters take
advantage of the new indenting and nesting functionality added to the
Simple formatter struct.

Fixes #163, fixes #203